Replaced ionic polyfills with babel-polyfill and manually including polyfills
